requirements.txt
Flask==2.3.3
Pillow==10.0.0
rembg==2.0.50
Werkzeug==2.3.7
gunicorn==21.2.0

